error container refactored
